Fix p4-print issue that cause lsp-mode and other packages problems
Override run-mode-hooks when fortifying p4-print buffers. This prevents errors from lsp-mode and
other packages that rely on hooks.

To setup lsp-mode you define a c++-mode-hook which will activate lsp-mode on a C++ buffer when
buffer-file-name is t (buffer is visiting a file). However, p4-print when p4 print'ing a file to
diff against creates a non-file buffer but sets buffer-file-name so we can fontify the buffer.  This
was causing problems for lsp-mode, so prevent the hook from running by overriding
run-mode-hooks. This will also improve performance.